requirements.txt
Django==4.0.2
django-simple-history==3.0.0
djangorestframework==3.13.1
Pillow==9.0.1